+++ src/app/(with-main-header)/Timer.tsx
@@ -1,38 +1,20 @@
-import TimerList from "./TimerList";
 import TimerCategory from "./TimerCategory";
 import { timerApi } from "@/api/timer/timer";
+import TimerList from "./TimerList";
+import { Suspense } from "react";
+import BasicSkeleton from "@/components/common/skeletons/BasicSkeleton";
 
 export default async function Timer() {
   const fetchTimerList = await timerApi.getTimerList();
   const timerList = fetchTimerList?.data;
-  const fetchStudyTimeForDaily = await timerApi.getStudyTimeForDaily();
-  const studyTimeForDaily = fetchStudyTimeForDaily?.data;
 
   if (!timerList) return;
-  if (!studyTimeForDaily) return;
-
   return (
     <section className="w-full lg:w-120 flex flex-col gap-4">
       <TimerCategory timerList={timerList} />
-      <div className="px-3 py-3 bg-site-button rounded-lg">
-        <div className="flex justify-center items-center gap-7 mt-2 lg:mt-3 mb-5 lg:mb-6">
-          <p className="w-34 text-[32px] text-site-darkgray-02 font-galmuri">
-            {studyTimeForDaily.studyTime}
-          </p>
-        </div>
-        <div className="flex flex-col gap-3">
-          {timerList.map((timer) => (
-            <TimerList
-              key={timer.timerId}
-              timerId={timer.timerId}
-              categoryId={timer.categoryId}
-              name={timer.name}
-              studyDate={timer.studyDate}
-              studyTime={timer.studyTime}
-            />
-          ))}
-        </div>
-      </div>
+      <Suspense fallback={<BasicSkeleton className="h-66 lg:h-76" count={1} />}>
+        <TimerList timerList={timerList} />
+      </Suspense>
     </section>
   );
 }
