+++ src/app/(with-main-header)/TodoItems.tsx
@@ -1,7 +1,9 @@
 import { todoApi } from "@/api/todo/todo";
 import TodoItem from "./TodoItem";
+import { delay } from "@/utils/delay";
 
 export default async function TodoItems() {
+  await delay(3000);
   const todos = await todoApi.getTodos();
   return (
     <div className="px-3 py-3 bg-site-button rounded-lg flex flex-col gap-3">
